DS-333: move translation config under translation service, change ATTENTION object position
